reworked email template payloads to have inheritance + more information
