+++ service/src/common/email-template-payload/base.journey.email.payload.ts
@@ -1,0 +1,10 @@
+import { JourneyType } from '../enums/journey.type';
+import { BaseEmailPayload } from './base.email.payload';
+
+export interface BaseJourneyEmailPayload extends BaseEmailPayload {
+  journey: {
+    name: string;
+    url: string;
+    type: JourneyType;
+  };
+}
