+++ service/src/common/email-template-payload/collaboration.card.created.email.payload.ts
@@ -1,20 +1,12 @@
-import { BaseEmailPayload } from './base.email.payload';
+import { BaseJourneyEmailPayload } from './base.journey.email.payload';
 
-export interface CollaborationCardCreatedEmailPayload extends BaseEmailPayload {
+export interface CollaborationCardCreatedEmailPayload
+  extends BaseJourneyEmailPayload {
   createdBy: {
     firstname: string;
     email: string;
   };
-  aspect: {
+  card: {
     displayName: string;
-  };
-  recipient: {
-    firstname: string;
-    email: string;
-    notificationPreferences: string;
-  };
-  community: {
-    name: string;
-    url: string;
   };
 }
