+++ service/src/common/email-template-payload/community.application.created.email.payload.ts
@@ -1,25 +1,11 @@
-import { BaseEmailPayload } from '@common/email-template-payload';
-import { JourneyType } from '../enums/journey.type';
+import { BaseJourneyEmailPayload } from './base.journey.email.payload';
 
 export interface CommunityApplicationCreatedEmailPayload
-  extends BaseEmailPayload {
+  extends BaseJourneyEmailPayload {
   applicant: {
     name: string;
     firstname: string;
     email: string;
     profile: string;
-  };
-  recipient: {
-    name: string;
-    email: string;
-    notificationPreferences: string;
-  };
-  community: {
-    name: string;
-    type: JourneyType;
-    url: string;
-  };
-  hub: {
-    url: string;
   };
 }
