+++ service/src/common/email-template-payload/community.new.member.email.payload.ts
@@ -1,17 +1,10 @@
-import { BaseEmailPayload } from './base.email.payload';
+import { BaseJourneyEmailPayload } from './base.journey.email.payload';
 
-export interface CommunityNewMemberEmailPayload extends BaseEmailPayload {
+export interface CommunityNewMemberEmailPayload
+  extends BaseJourneyEmailPayload {
   member: {
     email: string;
     name: string;
     profile: string;
-  };
-  recipient: {
-    firstname: string;
-    email: string;
-    notificationPreferences: string;
-  };
-  community: {
-    name: string;
   };
 }
