+++ service/src/common/email-template-payload/index.ts
@@ -1,4 +1,5 @@
 export * from './base.email.payload';
+export * from './base.journey.email.payload';
 export * from './community.application.created.email.payload';
 export * from './collaboration.card.comment.email.payload';
 export * from './collaboration.card.created.email.payload';
@@ -6,6 +7,6 @@ export * from './collaboration.callout.published.email.payload';
 export * from './communication.discussion.created.email.payload';
 export * from './communication.update.created.email.payload';
 export * from './collaboration.interest.email.payload';
-export * from './community.context.email.payload';
+export * from './collaboration.context.email.payload';
 export * from './community.new.member.email.payload';
 export * from './platform.user.registered.email.payload';
