+++ service/src/common/email-template-payload/platform.user.registered.email.payload.ts
@@ -7,12 +7,4 @@ export interface PlatformUserRegisteredEmailPayload extends BaseEmailPayload {
     email: string;
     profile: string;
   };
-  recipient: {
-    firstname: string;
-    email: string;
-    notificationPreferences: string;
-  };
-  hub: {
-    url: string;
-  };
 }
