+++ service/src/services/domain/builders/platform-user-registered/platform.user.registered.notification.builder.ts
@@ -79,11 +79,11 @@ export class PlatformUserRegisteredNotificationBuilder
         profile: registrantProfileURL,
       },
       recipient: {
-        firstname: recipient.firstName,
+        firstName: recipient.firstName,
         email: recipient.email,
         notificationPreferences: notificationPreferenceURL,
       },
-      hub: {
+      platform: {
         url: hubURL,
       },
     };
